First Working Reliable Broadcast Channel!!!!!!!!
:lit: :lit: :lit:

Now the client passes their message to the protocol layer.
This layer keeps track of all the individual message states.
The protocol layer passes messages down to the connection
pool as it sees fit

There is a state_manager.go file that manages what to do
when a particular state gets an incoming message